fix: Allow skipping aliased hooks
There are situations where a hook with the same ID will be used multiple times, but with
differing configurations. In the event that one of those configurations should not
desired (or can't) be run in pre-commit CI it should be possible to skip using the
`alias` configuration parameter. This adds support for the `alias` as an alternate to
just `id` in the `ci:skip` list.